+++ tests/integration/proxy/v1/service/test_session.py
@@ -32,7 +32,7 @@ class SessionTestCase(IntegrationTestCase):
             '''
             {
                 "service_sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "status": "in-progess",
+                "status": "in-progress",
                 "unique_name": "unique_name",
                 "date_started": "2015-07-30T20:00:00Z",
                 "date_ended": "2015-07-30T20:00:00Z",
@@ -113,7 +113,7 @@ class SessionTestCase(IntegrationTestCase):
             '''
             {
                 "service_sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "status": "in-progess",
+                "status": "in-progress",
                 "unique_name": "unique_name",
                 "date_started": "2015-07-30T20:00:00Z",
                 "date_ended": "2015-07-30T20:00:00Z",
@@ -181,7 +181,7 @@ class SessionTestCase(IntegrationTestCase):
             '''
             {
                 "service_sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "status": "in-progess",
+                "status": "in-progress",
                 "unique_name": "unique_name",
                 "date_started": "2015-07-30T20:00:00Z",
                 "date_ended": "2015-07-30T20:00:00Z",